Refactor JsEvaluator into focused evaluators
