+++ src/Asynkron.JsEngine/Ast/TypedAstEvaluator.cs
@@ -1312,7 +1312,7 @@ public static class TypedAstEvaluator
                     builder.Append(',');
                 }
 
-                builder.Append(JsEvaluator.ToJsStringForArray(array.Items[i]));
+                builder.Append(JsExpressionEvaluator.ToJsStringForArray(array.Items[i]));
             }
 
             return builder.ToString();
@@ -1342,7 +1342,7 @@ public static class TypedAstEvaluator
     {
         if (left is string || right is string || left is JsObject || left is JsArray || right is JsObject || right is JsArray)
         {
-            return JsEvaluator.ToJsString(left) + JsEvaluator.ToJsString(right);
+            return JsExpressionEvaluator.ToJsString(left) + JsExpressionEvaluator.ToJsString(right);
         }
 
         if (left is JsBigInt leftBigInt && right is JsBigInt rightBigInt)
